connected_stack/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyjwt==2.8.1
python-multipart==0.0.6
python-dotenv==1.0.0
